<!DOCTYPE html> <html lang="en"> <head>      <meta charset="UTF-8">     <meta name="viewport" content="width=device-width, initial-scale=1.0">      <link rel="stylesheet" href="https://unpkg.com/leaflet@1.8.0/dist/leaflet.css"/>     <link href="css/index.css" rel="stylesheet">     <script src="https://unpkg.com/leaflet@1.8.0/dist/leaflet.js" defer></script>     <script type="module" src="https://cdn.jsdelivr.net/npm/ionicons/dist/ionicons/ionicons.esm.js" defer></script>     <script nomodule src="https://cdn.jsdelivr.net/npm/ionicons/dist/ionicons/ionicons.js" defer></script>     <script src="final project.js" defer></script>     <title>Journey</title> </head> <body> <div class="container">     <header class="page-header">         <h1 class="page-title"> Fantastic Journey</h1>         <nav>             <ul class="flex-wrap space-evenly">                 <li><a class = "button button-white" href="final project.html">Index</a></li>                 <li><a class = "button button-white" href="about.html">About</a></li>             </ul>         </nav>     </header>     <main class="flex-wrap">         <header>             <h1>About the game</h1>         </header>         <article>             <header>                 <h5>We have completed the preliminary project in first period. There is game target for users. The game provides different options to users. For example, the user can choose some European capital cities to visit. There is feedback information from the system. The information was stored in the new-created table. We think that we achieved our aim of original design.<br><br>
Of course, there should be improvement and modifications in this game. We have discussed about the adjustments, meanwhile we also compare our project with the other groups projects. There are different targets in other groups. For instance: some projects would calculate the consumption of CO2, time zone and so on. We agree that there should be more quantity target in our project. We will add more data in further with usage of API. Just like we did in our assignment, we could use weather API, it provides instant information from  most of the big cities all around the world.<br><br>

In the second period, we will also learn HTML, CSS. We are thinking about how to combine the project with these new languages. But that is the preliminary idea. We did not decide how to complete it yet. In our opinion, it should be more complicated than the former one and provide more fantastic experience for users.
</h5>
            </header>
            <figure>

            </figure>
        </article>
    </main>
    <footer class="page-footer">
        <p>It is good to have an end to journey toward; but it is the journey that matters, in the end.”</p>
    </footer>
</div>
</body>
</html>